[Feat] Integration of showing all voting asset
