+++ frontend/src/types/index.ts
@@ -1,5 +1,13 @@
 // Frontend-related Types
 // Data view templates
+export interface GameAsset {
+    id: number;
+    image: string;
+    title: string;
+    proposer: string;
+    description: string;
+}
+
 export interface AssetResponse {
     creator: string;
     name: string;
